+++ lib/prom_ex/plugins/ecto.ex
@@ -18,6 +18,9 @@ if Code.ensure_loaded?(Ecto) do
        If you do not provide this value, PromEx will attempt to resolve your Repo modules via the
        `:ecto_repos` configuration on your OTP app.
 
+    - `duration_unit`: This is an OPTIONAL option and is a `Telemetry.Metrics.time_unit()`. It can be one of:
+      `:second | :millisecond | :microsecond | :nanosecond`. It is `:millisecond` by default.
+
     This plugin exposes the following metric groups:
     - `:ecto_init_event_metrics`
     - `:ecto_query_event_metrics`
@@ -44,6 +47,7 @@ if Code.ensure_loaded?(Ecto) do
     def event_metrics(opts) do
       otp_app = Keyword.fetch!(opts, :otp_app)
       metric_prefix = Keyword.get(opts, :metric_prefix, PromEx.metric_prefix(otp_app, :ecto))
+      duration_unit = Keyword.get(opts, :duration_unit, :millisecond)
 
       show_write_stats = Keyword.get(opts, :show_write_stats, true)
       show_read_stats = Keyword.get(opts, :show_read_stats, true)
@@ -72,7 +76,7 @@ if Code.ensure_loaded?(Ecto) do
       # Event metrics definitions
       [
         init_metrics(metric_prefix),
-        query_metrics(metric_prefix)
+        query_metrics(metric_prefix, duration_unit)
       ]
     end
 
@@ -171,13 +175,15 @@ if Code.ensure_loaded?(Ecto) do
       )
     end
 
-    defp query_metrics(metric_prefix) do
+    defp query_metrics(metric_prefix, duration_unit) do
+      duration_unit_plural = String.to_atom("#{duration_unit}s")
+
       Event.build(
         :ecto_query_event_metrics,
         [
           # Capture the db connection idle time
           distribution(
-            metric_prefix ++ [:repo, :query, :idle, :time, :milliseconds],
+            metric_prefix ++ [:repo, :query, :idle, :time, duration_unit_plural],
             event_name: @query_event,
             measurement: :idle_time,
             description: "The time the connection spent waiting before being checked out for the query.",
@@ -186,12 +192,12 @@ if Code.ensure_loaded?(Ecto) do
             reporter_options: [
               buckets: [10, 50, 250, 1_000, 5_000, 10_000]
             ],
-            unit: {:native, :millisecond}
+            unit: {:native, duration_unit}
           ),
 
           # Capture the db connection queue time
           distribution(
-            metric_prefix ++ [:repo, :query, :queue, :time, :milliseconds],
+            metric_prefix ++ [:repo, :query, :queue, :time, duration_unit_plural],
             event_name: @query_event,
             measurement: :queue_time,
             description: "The time spent waiting to check out a database connection.",
@@ -200,12 +206,12 @@ if Code.ensure_loaded?(Ecto) do
             reporter_options: [
               buckets: [10, 50, 250, 1_000, 5_000, 10_000]
             ],
-            unit: {:native, :millisecond}
+            unit: {:native, duration_unit}
           ),
 
           # Capture the db query decode time
           distribution(
-            metric_prefix ++ [:repo, :query, :decode, :time, :milliseconds],
+            metric_prefix ++ [:repo, :query, :decode, :time, duration_unit_plural],
             event_name: @query_event,
             measurement: :decode_time,
             description: "The time spent decoding the data received from the database.",
@@ -214,12 +220,12 @@ if Code.ensure_loaded?(Ecto) do
             reporter_options: [
               buckets: [5, 50, 100, 500, 2_500]
             ],
-            unit: {:native, :millisecond}
+            unit: {:native, duration_unit}
           ),
 
           # Capture the query execution time
           distribution(
-            metric_prefix ++ [:repo, :query, :execution, :time, :milliseconds],
+            metric_prefix ++ [:repo, :query, :execution, :time, duration_unit_plural],
             event_name: @query_event,
             measurement: :query_time,
             description: "The time spent executing the query.",
@@ -228,12 +234,12 @@ if Code.ensure_loaded?(Ecto) do
             reporter_options: [
               buckets: [10, 50, 250, 2_500, 10_000, 30_000]
             ],
-            unit: {:native, :millisecond}
+            unit: {:native, duration_unit}
           ),
 
           # Capture the total time (the sum of all other measurements)
           distribution(
-            metric_prefix ++ [:repo, :query, :total, :time, :milliseconds],
+            metric_prefix ++ [:repo, :query, :total, :time, duration_unit_plural],
             event_name: @query_event,
             measurement: :total_time,
             description: "The sum of the other time measurements.",
@@ -242,7 +248,7 @@ if Code.ensure_loaded?(Ecto) do
             reporter_options: [
               buckets: [10, 50, 250, 2_500, 10_000, 30_000]
             ],
-            unit: {:native, :millisecond}
+            unit: {:native, duration_unit}
           ),
 
           # Capture the number of results returned