+++ lib/prom_ex/utils.ex
@@ -7,6 +7,8 @@ defmodule PromEx.Utils do
   @typedoc "The kinds of exceptions that can occur"
   @type exception_kind :: :error | :exit | :throw
 
+  @type duration_unit_plural :: :seconds | :milliseconds | :microseconds | :nanoseconds
+
   @doc """
   Take a module name and normalize it for use as a metric
   label.
@@ -54,4 +56,13 @@ defmodule PromEx.Utils do
   def normalize_exception(_kind, _reason, _stacktrace) do
     "UnknownException"
   end
+
+  @doc """
+  Converts a `time_unit` to its plural form.
+  """
+  @spec make_plural_atom(System.time_unit()) :: atom()
+  def make_plural_atom(:second), do: :seconds
+  def make_plural_atom(:millisecond), do: :milliseconds
+  def make_plural_atom(:microsecond), do: :microseconds
+  def make_plural_atom(:nanosecond), do: :nanoseconds
 end